Correct predictors to deal with situations when zero is a bad guess (#17)
Change the prediction strategy when you're looking back into the history and don't have enough prior history to predict the next complete block.  Previously we would fill with zeros, now we will (1) use as much previous history as possible and then (2) use the zeroth entry in the history for the rest, rather than zeros.